Compiler: dont emit code for unused declarations
We now mark declarations as used when we use them (duh),
e.g. a function in a call expression. Then we don't emit llvm stuff
for those declarations, which greatly improves llvm performance.
Some functions need special treatment like the main function and
trait functions. The user can also add the #unused directive to a
function to tell the compiler to emit llvm code for this function
even if it is not used. ___chkstk_ms uses this.