Compiler: added enum operators and enum flags
you can now add the #flags directive to an enum. This means the enum is also #repr("C")
and the tag values will be powers of two, starting at 0 (0, 1, 2, 4, 8, ...), though you can
still provide an arbitrary value to each enum member. Enum flags also have the 'or' and 'and'
operators. 'or' is just the binary or of the tags, 'and' is a bool which is true if the binary and
of both tags is not 0.

Also added == and != to all enums.