Compiler: bug fixes
There was a bug where the type of a constant array of functions would be
the type of the first function (which is intended), but the type of the
first function had a reference to the declaration of this function, which
means the type of the array would now also have a reference to this function.
When indexing into that array and calling it the compiler saw that the array
had a reference to a function and called that function directly instead of
looking up the actual value in the array.

Now function types don't have a reference to the declaration anymore, it is
instead stored in the value of the function expression.